feat: add replay test reporters
